Updated the README.file to include assignment requirements
I made sure to include a description of my code, what it is supposed to output, and what the grader should expect to see so that it is clear and easy to understand. I provide simple steps that the grader needs to follow to be able to run my code. At the end, I include a paragraph consisting of my personal suggestions to how the assignment can be modified for the future.